Fixing various memory corruptions in src/graphics.c triggered by failed image loading leading to crashes

+++ src/audio.c
@@ -31,9 +31,9 @@ static void*(*samplecallbackptr)(unsigned int) = NULL;
 static int soundenabled = 0;
 
 void audiocallback(void *intentionally_unused, Uint8 *stream, int len) {
-	memset(stream, 0, (unsigned int)len);
+	memset(stream, 0, (size_t)len);
 	if (!samplecallbackptr) {return;}
-	SDL_MixAudio(stream, samplecallbackptr((unsigned int)len), (int)len, SDL_MIX_MAXVOLUME);
+	SDL_MixAudio(stream, samplecallbackptr((unsigned int)len), (unsigned int)len, SDL_MIX_MAXVOLUME);
 	//memcpy(stream, samplecallbackptr((unsigned int)len), (unsigned int)len);
 }
